# [PERF] Removes closures around `eachChar`