--- a/lib/route-recognizer.js
+++ b/lib/route-recognizer.js
@@ -30,13 +30,15 @@ function isArray(test) {
 
 function StaticSegment(string) { this.string = string; }
 StaticSegment.prototype = {
-  eachChar: function(callback) {
+  eachChar: function(currentState) {
     var string = this.string, ch;
 
     for (var i=0; i<string.length; i++) {
       ch = string.charAt(i);
-      callback({ validChars: ch });
+      currentState = currentState.put({ validChars: ch });
     }
+
+    return currentState;
   },
 
   regex: function() {
@@ -50,8 +52,8 @@ StaticSegment.prototype = {
 
 function DynamicSegment(name) { this.name = name; }
 DynamicSegment.prototype = {
-  eachChar: function(callback) {
-    callback({ invalidChars: "/", repeat: true });
+  eachChar: function(currentState) {
+    return currentState.put({ invalidChars: "/", repeat: true });
   },
 
   regex: function() {
@@ -65,8 +67,8 @@ DynamicSegment.prototype = {
 
 function StarSegment(name) { this.name = name; }
 StarSegment.prototype = {
-  eachChar: function(callback) {
-    callback({ invalidChars: "", repeat: true });
+  eachChar: function(currentState) {
+    return currentState.put({ invalidChars: "", repeat: true });
   },
 
   regex: function() {
@@ -80,7 +82,9 @@ StarSegment.prototype = {
 
 function EpsilonSegment() {}
 EpsilonSegment.prototype = {
-  eachChar: function() {},
+  eachChar: function(currentState) {
+    return currentState;
+  },
   regex: function() { return ""; },
   generate: function() { return ""; }
 };
@@ -307,14 +311,6 @@ function findHandler(state, path, queryParams) {
   return result;
 }
 
-function addSegment(currentState, segment) {
-  segment.eachChar(function(ch) {
-    currentState = currentState.put(ch);
-  });
-
-  return currentState;
-}
-
 function decodeQueryParamPart(part) {
   // http://www.w3.org/TR/html401/interact/forms.html#h-17.13.4.1
   part = part.replace(/\+/gm, '%20');
@@ -360,7 +356,7 @@ RouteRecognizer.prototype = {
         regex += "/";
 
         // Add a representation of the segment to the NFA and regex
-        currentState = addSegment(currentState, segment);
+        currentState = segment.eachChar(currentState);
         regex += segment.regex();
       }