Instead of closing over arguments we can directly call the function with
currentState. Does not include any of the other PRs.

```
  Add ................... 606.20 op/s
  End-to-end ............ 283.64 op/s
  Generate ........ 4,872,032.98 op/s
  Handlers For ... 39,028,332.01 op/s
  Recognize ......... 366,043.01 op/s
```