@stefanpenner 2016-01-29
@chadhietala sharing before/after is valuable, sharing a single set of numbers isn't really.