@krisselden 2016-01-29
@stefanpenner agreed, he was referencing the numbers posted in the first PR which did have 2 sets but that will be lost looking back at the log.

https://github.com/tildeio/route-recognizer/pull/75#issuecomment-176510303