@krisselden 2016-01-29
@stefanpenner sorry, wrong PR, it was the PR with the benchmarks added that had the baseline https://github.com/tildeio/route-recognizer/pull/76